[Refactor] (code) Move the objects to their own single modules:
*XXXSGICmd* moves to '.server.sgi.cmd'.

*XXXSGICmdOption* moves to '.server.sgi.cmdoption'.
