+++ pymock_api/server/sgi/__init__.py
@@ -1,0 +1,4 @@
+"""*The wrapper of application which ruled by SGI (server gateway interface)*
+
+The factory to generate SGI application instance to run Python web framework.
+"""
